+++ src/main/java/com/money/SaveMi/Service/CurrencyService.java
@@ -1,6 +1,7 @@
 package com.money.SaveMi.Service;
 
 import com.money.SaveMi.DTO.Currency.SaveCurrencyDto;
+import com.money.SaveMi.DTO.Shared.BulkDeleteDto;
 import com.money.SaveMi.Model.Currency;
 import com.money.SaveMi.Model.User;
 import com.money.SaveMi.Repo.CurrencyRepo;
@@ -12,40 +13,53 @@ import org.springframework.stereotype.Service;
 public class CurrencyService {
     private final CurrencyRepo currencyRepo;
     private final UserRepo userRepo;
-    private final AuthenticationServiceUtil authenticationServiceUtil;
+    private final AuthenticationServiceUtil authUtil;
 
-    public CurrencyService(CurrencyRepo currencyRepo, UserRepo userRepo, AuthenticationServiceUtil authenticationServiceUtil) {
+    public CurrencyService(CurrencyRepo currencyRepo, UserRepo userRepo, AuthenticationServiceUtil authUtil) {
         this.currencyRepo = currencyRepo;
         this.userRepo = userRepo;
-        this.authenticationServiceUtil = authenticationServiceUtil;
+        this.authUtil = authUtil;
     }
 
     public Iterable<Currency> getAllCurrenciesFromUser() {
-        String userUUID = authenticationServiceUtil.getCurrentUserUuid();
+        String userUUID = authUtil.getCurrentUserUuid();
         return currencyRepo.findAllCurrenciesByUserId(userUUID);
     }
 
     public Currency getCurrencyById(Long id) {
-        String userUUID = authenticationServiceUtil.getCurrentUserUuid();
+        String userUUID = authUtil.getCurrentUserUuid();
         return currencyRepo.findCurrencyByIdAndUserId(id, userUUID)
                 .orElseThrow(() -> new RuntimeException("Currency not found with id: " + id + " for user: " + userUUID));
     }
 
     public Currency saveCurrency(SaveCurrencyDto currency) {
-        String userUUID = authenticationServiceUtil.getCurrentUserUuid();
+        String userUUID = authUtil.getCurrentUserUuid();
         User user = userRepo.findById(userUUID)
                 .orElseThrow(() -> new IllegalArgumentException("User not found with UUID: " + userUUID));
         Currency newCurrency = new Currency(user, currency.name(), currency.symbol());
         return currencyRepo.save(newCurrency);
     }
 
-    public void deleteCurrency(Long id) {
-        String userUUID = authenticationServiceUtil.getCurrentUserUuid();
-        currencyRepo.deleteCurrencyByIdAndUserId(id,userUUID);
+    public void bulkDelete(BulkDeleteDto bulkdeleteDto){
+        String userId = authUtil.getCurrentUserUuid();
+
+        bulkdeleteDto.ids().forEach(id -> {
+            if(currencyRepo.findCurrencyByIdAndUserId(id,userId).isEmpty()){
+                throw new RuntimeException("Currency not found with id: " + id + " for user: " + userId + " in bulk");
+            }
+        });
+
+        currencyRepo.bulkDelete(bulkdeleteDto.ids(),userId);
     }
 
-    public boolean currencyExists(Long id) {
-        return currencyRepo.existsById(id);
+    public void deleteCurrency(Long id) {
+        String userUUID = authUtil.getCurrentUserUuid();
+
+        if(currencyRepo.findCurrencyByIdAndUserId(id,userUUID).isEmpty()){
+            throw new RuntimeException("Currency not found with id: " + id + " for user: " + userUUID);
+        }
+
+        currencyRepo.deleteCurrencyByIdAndUserId(id,userUUID);
     }
 
     public Currency findByNameAndSymbolAndUserId(String name, String symbol, String userId) {
